Fix repeated application of fixup commit changes
The list of commits-to-fixup would sometimes contain the same commit
twice, and for the same reason, be missing other commits.
This resulted in a spurious error message, after which the user had to
invoke git-autofixup again to fixup missing commits.

Here is a simple reproducer:

	git init tmp_repo
	cd tmp_repo
	git commit --allow-empty -m initial\ commit

	seq 1 20 > a
	git add a
	git commit -m A

	seq 1 10 > b
	git add b
	git commit -m B

	sed -i -e 2cX -e 18cX a
	sed -i -e 2cX b
	../git-autofixup HEAD~2

git-autofxup HEAD~2 before the fix:

	master 2b1471a] fixup! A
	 1 file changed, 2 insertions(+), 2 deletions(-)
	error: patch failed: a:1
	error: a: patch does not apply
	error: patch failed: a:15
	error: a: patch does not apply
	git apply: non-zero exit code

git-autofxup HEAD~2 after the fix:

	[master b7e0124] fixup! A
	 1 file changed, 2 insertions(+), 2 deletions(-)
	[master 1810461] fixup! B
	 1 file changed, 1 insertion(+), 1 deletion(-)