Only consider staged hunks, if there are any
Instead of aborting when anything is staged, automatically choose
whether to consider all hunks or staged hunks, based on whether any
hunks are staged.

This obsoletes the --use-index option. A --dry-run option option has
been added, because it basically fell out of the changes for this.

It seems better for maintainability and testing to have a single way to
do things, so always use a temp git index. Simplifying the interface by
getting rid of --use-index was also nice.

Fixes #5