Only use temp index if hunks are staged
For big repos reading HEAD's tree into the temp index can apparently
take a while, so we should skip creating it if we can. Also, if we use a
temp index for unstaged hunks, then the original index will be
out-of-sync with any hunks that were committed via the temp index.

If you hit this, `git restore --staged` or `git read-tree HEAD^{tree}`
can be used to read the new HEAD's tree into the index.

In tests, if no staged changes are explicitly expected via the
`staged_want` param, fail if anything is in the index after calling
main().

Fixes #16