Remove README.pod
README.pod is generated from the POD in git-autofixup, so just make it
when needed for creating release tarballs. I don't think there's a need
to keep it around otherwise. Keeping it can be confusing because it's
not obvious that it's generated, and getting rid of it is easier than
writing a pre-commit hook to keep it up to date.