Choose upstream revisions automatically
If an upstream revision isn't given on the command-line, try to find
suitable ones automatically. This only works if the current branch has a
tracking branch.

Writing tests for this feature required refactoring the existing test
helpers significantly, and it made sense to put things in different
files and use packages other than main to make it obvious what's being
called. The stub module has to use the package that the module is
published as on CPAN, and we want to put the version in the script
(git-autofixup) as well and compare the two, so the script needs to be a
different package, which will only be used internally.

Closes #27