Improve error handling for external/system commands
For external commands that are expected to fail, it seems nice to print
their stderr, annotated with the command so it doesn't seem like the
error message is from git-autofixup itself. So capture() has been added
to make this easier to do.

For external commands that are unlikely to fail, it's still nice to
check for failure to avoid continuing in a confusing way.

Previously I was using $! when it wasn't meaningful, often when I should
have been using $? instead.

Also, give more detailed information about problems with external
commands, adding child_error_desc() to make it more convenient.