Load Pod::Usage at runtime since Git for Windows doesn't include it
Pod::Usage is in the standard library, but Git for Windows includes HTML
documentation for git commands and does special handling of --help
options to open the HTML docs in a browser, so I guess they thought it
wouldn't be necessary.

It doesn't seem worth it to make HTML docs and installation instructions
for Git for Windows.