Fix some cleanup issues in repo.pl
I was getting warnings when running tests via `make test` that I wasn't
seeing otherwise, and realized that I wasn't using strict and warnings.

Looks like I was accidentally turning off cleanup for repo tempdirs
since I was missing the $ sigil on self in _init_env().

And the cleanup logic for env vars had a few issues. If a variable
didn't exist in the original env, better to delete it than to set it to
undef. And I was missing the $ sigil for key when getting the original
$val.