Don't fatalize warnings
`perldoc warnings` discourages using `FATAL => 'all'` since it hurts
portability and goes against the expectations of library writers, who
expect warnings not to be fatal. If Perl adds some warning in the
future, a program that's fatalizing all warnings could stop working. And
yeah, it's better for a user to get a warning than for the program to
not work.

It seems fine to fatalize warnings for test code, though.

Of course, I'd still like to have git-autofixup produce as few warnings
as possible, but I'm not sure it's worth using Test::NoWarnings. Perhaps
being careful will be enough.