Handle side-prefixes (a/, b/, etc) in diff output better
Back in 2019 (commit 2e199745ae0) we started using `-c
diff.noprefix=false` so that we could keep getting the default a/ and b/
prefixes regardless of the configured diff.noprefix value, and later we
started stripping the diff.mnemonicPrefix prefiexes, too. Now
diff.srcPrefix and diff.dstPrefix have been added to git to make the
prefixes user-configurable, and it seems preferable to solve the issue
of handling these diff side-prefixes more generally rather than try to
plug one more hole.

The motivating issue at the moment is that if diff.dstPrefix is set to
"./" we hit an error like the following when trying to apply changes to
the git index. Apparently from the perspective of `git apply`, "a" and
"./a" are distinct filenames.

    error: ./a: does not exist in index

Diff output is consumed by two things in git-autofixup: parse_hunks()
and git apply, and both need to know what prefix to strip. `git apply`
works kinda like the `patch` command in this regard and by default
removes the leading path component from filenames in hunk headers and
can optionally remove any number of components. The `-p` option for `git
apply` must be set appropriately if the prefix isn't a single path
component, like if we're using diff.noPrefix or diff.dstPrefix. It'd be
nice if we could use git diff's --default-prefix option to override all
the `diff.*Prefix` config, but it was only added in git 2.41, and I'd
like to keep backwards compability when feasible. `--no-prefix` has been
around since Git 1.5.4, and should work fine as long as we set `-p`
appropriately for git apply and stop stripping prefixes in
dequote_diff_filename() (called by parse_hunks()). Prior to git commit
6a2abdc125 (released in v1.7.11), `-p0` didn't work right for binary
diffs, but based on the commit message it sounds like it still worked
for text diffs, and all the tests pass with 1.7.4.

Testing:

- remove side prefixes from tests and update the commands in Repo for
  getting diffs for tests
- add script to test with old versions of git, built from source
- avoid using `rebase -q` and redirect output instead, since `-q` didn't
  exist in 1.7.4
- skip test that uses GIT_SEQUENCE_EDITOR, which didn't exist until
  1.7.8

Fixes #31